#50 #51 #52 Feat: 관리자 웨이팅 관리 front & back 구현 완료

+++ src/main/java/kr/oshino/eataku/waiting/model/dto/requestDto/ReadWaitingRequestDto.java
@@ -11,6 +11,6 @@ import lombok.*;
 public class ReadWaitingRequestDto {
 
     private Long memberNo;
-    private Integer restaurantNo;
+    private Long restaurantNo;
     private StatusType waitingStatus;       // 예약 전체 정보 or 대기중 정보 분리
 }
